requirements.txt
face_recognition
opencv-python
numpy
pillow 
tkinter 
pyinstaller
setuptools